% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRstoAAC.R
\name{byGene}
\alias{byGene}
\title{byGene}
\usage{
byGene(
  genes,
  gencode_basic_uniprot,
  edb,
  ccr,
  outfile,
  chr,
  fastafolder,
  ens_version,
  np
)
}
\arguments{
\item{genes}{Your list of genes}

\item{gencode_basic_uniprot}{Mapping table, Ensembl and UniProtKB identifiers will be taken from here}

\item{edb}{a ensembldb database connection}

\item{ccr}{the CCRs dataframe with sites, CCRs percentiles, etc}

\item{outfile}{the output file}

\item{chr}{chromosome of interest}

\item{fastafolder}{will check if the protein sequences are already here, if not donwload}

\item{ens_version}{version of Ensembl}

\item{np}{number of proceses to run in parallel}
}
\value{
A dataframe with final mapping
}
\description{
Goes gene-by-gene in your list, calling the mapping of positions
This function will do parallel calls to 'np' processors, dividing 'bysite' in chunks
among them
}
