% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRStoAAC.R
\name{CCRStoAAC}
\alias{CCRStoAAC}
\title{CCRStoAAC}
\usage{
CCRStoAAC(
  gene = NULL,
  chromosome = NULL,
  nproc = NULL,
  gnom_version = "gnomAD3_0",
  ens_version = 101,
  out_path = "out/",
  ccr_file =
    "data/rawCCRs/gnomad3_0/vep101/sort_weightedresiduals-cpg-synonymous-novariant.txt.gz",
  mapping_file = "data/mapping_tables/ensembl_uniprot_MANE_metrics_07102020.tsv.gz",
  gtf = "data/GTF101/",
  fastas = "data/fastas",
  keep = 0
)
}
\arguments{
\item{gene}{Genes of interest, comma separated. Passing "all" (default) triggers the mapping of all genes in a chromosome}

\item{chromosome}{Only autosomes and X chromosome. Mandatory parameter.}

\item{nproc}{Number of cores for parallel processing}

\item{gnom_version}{gnomAD database version. Default is "gnomad3_0". It will be used for naming output files}

\item{ens_version}{Ensembl database version employed for annotating gnomAD VCF files. Will define the set of transcripts for the #' #' mapping}

\item{out_path}{Path to output folder}

\item{ccr_file}{Path to the raw CCRs file, obtained after running the CCRs pipeline}

\item{mapping_file}{Path to the file with the mapping of IDs bewteen UniProtKB, Ensembl, MANE and gnomAD constraint scores}

\item{fastas}{Path to the folder where protein fasta files are/will be organized by chromosome (will be extended with 'chr')}

\item{keep}{keep=1 keeps previous output and resume after last mapped gene, keep=0 overwrites}

\item{gtf_file}{Path to folder where Ensembl GTF files are, splitted by chromosome for an easier lookup}
}
\value{
The function itself calls to do the mapping and saves the results to a file, returning nothing at it's end,
this is because the mapping tables can be quite large for a whole chromosome for example
}
\description{
This code is for mapping the Human Constrained Coding Regions (CCRs) from genomic
coordinates (chromosomes 1-22 and X) to amino acid sites in protein sequences
of UniProtKB, via reference annotated Ensembl transcripts of the GENCODE basic Project.

The parameters are set to default in order to demonstrate the mapping of CCRs for 3 genes of chromosome 18,
using as input the files organised in different sub-folders in the 'data/' directory of this package

External dependencies (non R libraries):
blastp
awk
grep
tr
sort
}
\examples{

 CCRStoAAC(gene="VPS4B",chromosome=18)
 CCRStoAAC( gene="all",chromosome=18,out_path="out/",nproc=10)
 CCRStoAAC( gene="WDR7,VAPA,VPS4B", chromosome=18,nproc=NULL, gnom_version="gnomAD3_0",
 ens_version=101,out_path="out/",ccr_file="data/rawCCRs/gnomad3_0/vep101/sort_weightedresiduals-cpg-synonymous-novariant.txt.gz",
 mapping_file="data/mapping_tables/ensembl_uniprot_MANE_metrics_07102020.tsv.gz",gtf="data/GTF101/",fastas="data/fastas",keep=0)

}
