% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRStoAAC.R
\name{getEnsDB}
\alias{getEnsDB}
\title{getEnsDB}
\usage{
getEnsDB(ens_version, chr)
}
\arguments{
\item{ens_version}{Version of Ensembl}

\item{chr}{Chromosome}
}
\value{
A data frame
}
\description{
Gets annotations of a specific version of Ensembl DB by using 'ensembldb'
Bioconductor R Package
Provide version of ensembl of your preference, but please use the same one used when
annotating gnomAD VCF file with VEP when running the CCRs pipeline
(ideally based on GRCh38/hg38) otherwise you might find discrepancies on the mapping!
}
