% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRstoAAC.R
\name{getENSPseq}
\alias{getENSPseq}
\title{getENSPseq}
\usage{
getENSPseq(ensembl_protein_id, file)
}
\arguments{
\item{ensembl_protein_id}{The Ensembl identifier (ENSP) of a protein sequence
to download in fasta format}

\item{file}{TThe location of the file where to save the *.fasta sequence}
}
\value{
Nothing
}
\description{
Downloads the Ensembl protein sequence for a ENSP identifier and
saves in a specified file
}
