% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRstoAAC.R
\name{getIDs}
\alias{getIDs}
\title{getIDs}
\usage{
getIDs(mapping_file, gtf_file, chr, ens_version)
}
\arguments{
\item{mapping_file}{The location of the gtf file for the actual chromosome}

\item{gtf_file}{The location of the gtf file for the actual chromosome}

\item{chr}{Chromosome}

\item{ens_version}{Ensembl version}
}
\value{
A data frame with identifiers
}
\description{
Gets the Ensembl-Uniprot IDs mapping, for those transcripts
that encode for proteins and that are part of the Gencode basic set of transcripts
}
