% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRStoAAC.R
\name{getRawCCRs}
\alias{getRawCCRs}
\title{getRawCCRs}
\usage{
getRawCCRs(ccr_file, chr = NULL, gene = NULL)
}
\arguments{
\item{ccr_file}{The location of the raw CCRs file, obtained from running the model}

\item{chr}{Chromosome}

\item{gene}{Gene}
}
\value{
A data frame with CCRs with the regions extended to each genomic base inside its boundaries
}
\description{
Reads in the CCRs file
Make sure the CCRs file is sorted by chrom and genomic position, this will further
simplify loading EnsDb.Hsapiens.v## in memory, doing it only once per chr
}
