% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRstoAAC.R
\name{getUniPseq}
\alias{getUniPseq}
\title{getUniPseq}
\usage{
getUniPseq(uniprot_blastp_id, gencode_basic_uniprot, file)
}
\arguments{
\item{uniprot_blastp_id}{The UniProtKB identifier of a protein sequence
to download in fasta format}

\item{gencode_basic_uniprot}{Mapping table, related identifiers will be taken from here}

\item{file}{The location of the file where to save the *.fasta sequence}
}
\value{
Nothing
}
\description{
Downloads the UniProtKB protein sequence for a given identifier and
saves in a specified file
}
