% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRstoAAC.R
\name{mapping}
\alias{mapping}
\title{mapping}
\usage{
mapping(
  gnm,
  bysite,
  selectedtx,
  gencode_basic_uniprot,
  blast_matches_seq,
  fastafolder,
  ens_version,
  chr
)
}
\arguments{
\item{gnm}{database connection, generated with with 'ensembldb'}

\item{bysite}{a dataframe with CCRs genomic regions transformed into one genomic
position by line}

\item{selectedtx}{the selected transcripts for the mapping}

\item{gencode_basic_uniprot}{Mapping table, related identifiers will be taken from here}

\item{fastafolder}{The location of the  *.fasta sequences}

\item{ens_version}{version of Ensembl}

\item{chr}{chromosome of interest}

\item{uniprot_blastp_id}{The UniProtKB identifier of a protein sequence
to download in fasta format}
}
\value{
A dataframe with mapping of coordinates
}
\description{
Performs the mapping of genomic positions to ensembl protein
}
