% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRstoAAC.R
\name{matchedProteinSeqs}
\alias{matchedProteinSeqs}
\title{matchedProteinSeqs}
\usage{
matchedProteinSeqs(
  ensembl_gene_ids,
  gencode_basic_uniprot,
  fastafolder,
  remove
)
}
\arguments{
\item{ensembl_gene_ids}{The UniProtKB identifier of a protein sequence
to download in fasta format}

\item{gencode_basic_uniprot}{Mapping table, related identifiers will be taken from here}

\item{fastafolder}{will check if the protein sequences are already here, if not donwload}

\item{ens_version}{version of Ensembl}
}
\value{
A dataframe
}
\description{
Gets the amino acid sequences and correspondences between ENSP and UniProt
Ensembl and UniProtKB proteins may differ in sequence even when they are assigned
as corresponding in these databases
This function runs blastp for comparing Ensembl and UniProtKB protein sequences
and makes sure they are 100\% identical
}
