% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRStoAAC.R
\name{byGene}
\alias{byGene}
\title{byGene}
\usage{
byGene(
  genes,
  gencode_basic_uniprot,
  edb,
  ccr,
  outfile,
  chr,
  fastafolder,
  ens_version,
  np,
  logfile
)
}
\arguments{
\item{genes}{Your full list of genes}

\item{gencode_basic_uniprot}{Mapping table, Ensembl and UniProtKB identifiers will be taken from here}

\item{edb}{a ensembldb database connection}

\item{ccr}{the CCRs dataframe with sites, CCRs percentiles, etc}

\item{outfile}{the output file}

\item{chr}{The corresponding chromosome of the genes}

\item{fastafolder}{will check if the protein sequences are already here, if not donwload}

\item{ens_version}{version of Ensembl}

\item{np}{number of proceses to run in parallel}

\item{logfile}{logfile for the run, default is "/dev/null"}
}
\value{
A dataframe with final mapping
}
\description{
Goes gene-by-gene in your list, calling the mapping of positions
This function will do parallel calls to 'np' processors, dividing 'bysite' in chunks
among them. Finally, it will dump the extended resulting mappings to all GENCODE basic
transcripts, and return a file with only the mappings for UniProt/SP canonical sequences
}
